==> 01b - Corpus statistics-Copy1.ipynb <==
# Retrieve content

from collections import namedtuple, Counter
import numpy as np

Dataset = namedtuple("Dataset", ["lemmas", "forms"])
Samples = namedtuple("Samples", ["train", "dev", "test"])
Corpus = namedtuple("Corpus", ["dataset", "samples"])

def get_lemmas(file: str = "./dataset/main/dev.txt"):
    out = {
        "positive": Dataset([], []),
        "negative": Dataset([], [])
    }
    
    with open(file) as f:
        for idx, orig_line in enumerate(f):
            if idx == 0:
                headers = orig_line.strip().split("\t")[1:]
                continue
            line = orig_line.split("\t")
            if len(line) != len(headers):
                if orig_line.startswith("#[TAG]"):
                    cat = orig_line.strip()[6:]
                continue
            line = dict(zip(headers, line))
            out[cat].lemmas.append(line["lemma"])
            out[cat].forms.append(line["token"])
    return out

def get_corpus(folder="./dataset/main/"):
    Global = {
        "positive": Dataset([], []),
        "negative": Dataset([], []),
    }
    samples = {
        "positive": [0, 0, 0],
        "negative": [0, 0, 0]
    }
    for idx, file in enumerate([f"{folder}train.txt", f"{folder}dev.txt", f"{folder}test.txt"]):
        current_output = get_lemmas(file)
        for key, dataset in current_output.items():
            for i in range(len(dataset)):
                Global[key][i].extend(dataset[i])
        
        with open(file) as f:
            for orig_line in f:
                if orig_line.startswith("#[TAG]"):
                    cat = orig_line.strip()[6:]
                    samples[cat][idx] += 1
    return Corpus(Global, {key: Samples(*val) for key, val in samples.items()})

Main = get_corpus()

# MLTD and Other metrics function

# from https://github.com/kristopherkyle/lexical_diversity/blob/master/lexical_diversity/lex_div.py


def safe_divide(numerator, denominator):
    if denominator == 0 or denominator == 0.0:
        index = 0
    else: index = numerator/denominator
    return index


def ttr(text):
    ntokens = len(text)
    ntypes = len(set(text))
    
    return safe_divide(ntypes,ntokens)

def mtld(input, min = 10): #original MTLD described in Jarvis & McCarthy 
    def mtlder(text):
        factor = 0
        factor_lengths = 0
        start = 0
        for x in range(len(text)):
            factor_text = text[start:x+1]
            if x+1 == len(text):
                factor += safe_divide((1 - ttr(factor_text)),(1 - .72))
                factor_lengths += len(factor_text)
            else:
                if ttr(factor_text) < .720 and len(factor_text) >= min:
                    factor += 1
                    factor_lengths += len(factor_text)
                    start = x+1
                else:
                    continue

        mtld = safe_divide(factor_lengths,factor)
        return mtld
    input_reversed = list(reversed(input))
    mtld_full = safe_divide((mtlder(input)+mtlder(input_reversed)),2)
    return mtld_full

# Our own code
def partage(k, j):
    return len(set(k).intersection(set(j))) / len(set(k)) * 100

def distribution(data):
    x = list(Counter(data).values())
    return np.std(x)

## Déviation standard

def get_pos_negs(theCorpus):
    positives = {
        "Taille": len(theCorpus.dataset["positive"].forms), 
        "Richesse": len(set(theCorpus.dataset["positive"].forms)),
        "MTLD(Lemmes)": mtld(theCorpus.dataset["positive"].lemmas), 
        "MTLD(Formes)": mtld(theCorpus.dataset["positive"].forms), 
        "Distribution(Lemmes)": distribution(theCorpus.dataset["positive"].lemmas),
        "Distribution(Formes)": distribution(theCorpus.dataset["positive"].forms),
        "Partage(Lemmes)": partage(theCorpus.dataset["positive"].lemmas, theCorpus.dataset["negative"].lemmas),
        "Partage(Formes)": partage(theCorpus.dataset["positive"].forms, theCorpus.dataset["negative"].forms),
        "Documents": theCorpus.samples["positive"]
    }
    negatives = {
        "Taille": len(theCorpus.dataset["negative"].forms), 
        "Richesse": len(set(theCorpus.dataset["negative"].forms)),
        "MTLD(Lemmes)": mtld(theCorpus.dataset["negative"].lemmas), 
        "MTLD(Formes)": mtld(theCorpus.dataset["negative"].forms), 
        "Distribution(Lemmes)": distribution(theCorpus.dataset["negative"].lemmas),
        "Distribution(Formes)": distribution(theCorpus.dataset["negative"].forms),
        "Partage(Lemmes)": partage(theCorpus.dataset["negative"].lemmas, theCorpus.dataset["positive"].lemmas),
        "Partage(Formes)": partage(theCorpus.dataset["negative"].forms, theCorpus.dataset["positive"].forms),
        "Documents": theCorpus.samples["negative"]
    }
    return positives, negatives

from pandas import DataFrame
df_main = DataFrame([*get_pos_negs(Main)], index=["Positifs", "Négatifs"]).transpose()
print(df_main.to_latex(float_format=lambda x: f"{x:.2f}"))

Partial = get_corpus("./dataset/main-partial/")
Metaphors = get_corpus("./dataset/metaphors/")
Literals = get_corpus("./dataset/inversed-metaphors/")
Literals = get_corpus("./dataset/inversed-metaphors/")

df_partial = DataFrame([*get_pos_negs(Partial)], index=["Positifs", "Négatifs"]).transpose()
print(df_partial.to_latex(float_format=lambda x: f"{x:.2f}"))


df_partial.drop(labels="Documents") - df_main.drop(labels="Documents")

print(Partial.samples)
print(Main.samples)
print(Literals.samples)
print(Metaphors.samples)
from pandas import MultiIndex

indexes= MultiIndex.from_tuples([
    (corpus, ds)
    for corpus in ["Général", "Partiel", "Litéral", "Métaphores"]
    for ds in ["Positif", "Négatif"]
], names=["Corpus", "Set"])

df = DataFrame([
    Main.samples["positive"],
    Main.samples["negative"],
    Partial.samples["positive"],
    Partial.samples["negative"],
    Literals.samples["positive"],
    Literals.samples["negative"],
    Metaphors.samples["positive"],
    Metaphors.samples["negative"],
    #Partial.samples,
    #Literals.samples,
    #Metaphors.samples
], index=indexes).sort_index(level="Set")

print(df.to_latex())
