==> 3 - Train and save specific models - Debugging Triplet.ipynb <==
from seligator.main import train_and_get, Seligator
from seligator.common.load_save import load
from seligator.prediction.tests import run_tests

params = load("./seligator/siamese.json")
params["reader_kwargs"]["batch_size"] = 16
params["reader_kwargs"]["dev_batch_size"] = 32
#params["batches_per_epoch"] = 5


seligator, reader, train, dev = Seligator.init_from_params(**params)
train_kwargs = dict(patience=4, num_epochs=50, lr=1e-3, optimizer="AdamW")
print(seligator.model)

model = train_and_get(seligator.model, train, dev, **train_kwargs)
data, img = run_tests(
    "dataset/main/test.txt",
    dataset_reader=reader, model=model, dump=f"siamese.miner.best.csv"
)
#seligator.save_model("siamese_saved_model")
