==> 04a - Visualize test results - Attention.ipynb <==
import pandas
import numpy as np
from colour import Color

has_attention = True
dont_show_ok = False
data = pandas.read_csv("./test_dumps/main.AggloMSD-HAN-None-Word2Vec-256.1.best.csv.gz")
# print(data.head)
data["attention"] = data["attention"].apply(lambda x: np.array(eval(x)))
data["ok"] = data["prediction"] == data["label"]



data

from IPython.core.display import display as _display, HTML  as _HTML
data.iloc[0]

COLORS = [Color("white")] + list(Color("white").range_to(Color("red"),50))

def visualize(sentence, attention, tag="p", colors=None, display=True, label=None, is_right: bool = False,
             show_score: bool = False, use_subword=False):
    if not colors:
        colors = COLORS

    # print(attention)
    
    if use_subword:
        sentence = sentence["token_subword"]
    else:
        sentence = sentence["sentence"]
        
    html = f"<{tag}>"
    if label:
        html += f"<b style=\"{'text-decoration: line-through;' if not is_right else ''};\">{label}</b> -> "
    if show_score:
        html += f"<i>{max(attention):.2f}</i> -> "
    
    if attention is not None:
        spans = [
            f"<span style=\"background-color: {colors[min(max(0, round(s*100) - 1), 50)]}; border-radius:5px; padding:2px; display: inline-block;\">{w}</span>"
            for w, s in zip(sentence.split(" "), attention.tolist())
        ]
    else:
        spans = [
            f"<span style=\"display: inline-block;\">{w}</span>"
            for w in sentence.split(" ")
        ]
        
    html += " ".join(spans) + f"</{tag}>"

    if display:
        _display(_HTML(html))
    return html

html = "<div>"
for idx, row in data.iterrows():
    if not(row["ok"] != True and row.prediction != "negative"):
        continue
    html += "    %s\n" % visualize(
        row,
        row.get("attention") if has_attention else None,
        label=row["label"],
        is_right=row["ok"],
        use_subword=False
    )
html += "</div>"
#print(html)
