==> 07b - Convert Bert Main for General evaluation.ipynb <==
import glob
import pandas as pd
import csv

merged_files = []
counter = {
    
}
for file in glob.glob("bert-experiment/bert_experiment*.csv"):
    with open(file) as f:
        r = csv.DictReader(f)
        for row in r:
            merged_files.append(row)
            if row["bert_trainable"] == "False" and row["bert_layer"] == "-1" and row.get("linear_enriched") == "True":
                if row["bert_mode"] in ["BertPoolerClass.HAN", "BertPoolerClass.MEANMAX", "BertPoolerClass.GRU"]:
                    #print(row)
                    mode = row["bert_mode"].split(".")[-1].replace("MEANMAX", "MeanMax")
                    row["name"] = f"NoMorph-{mode}-LinearEnriched-Bert-256"
                    if row["name"] not in counter:
                        counter[row["name"]] = 0
                    else:
                        counter[row["name"]] += 1
                    with open(f"csv_dumps/main.NoMorph.{mode}.LinearEnriched.Bert.256.{counter[row['name']]}.csv", "w") as f2:
                        w = csv.DictWriter(f2, fieldnames=row.keys())
                        w.writeheader()
                        w.writerow(row)
                    #print(row["bert_mode"], row["precision-positive"], row["recall-positive"], row["fscore-positive"])
                    #print(row)
            break
